+++ includes/DrupalSettingsHandler.php
@@ -10,8 +10,18 @@ class DrupalSettingsHandler {
 
   const VERSION_9 = 8;
 
-  function __construct(string $app_root, string $database_key = 'default', string $path_to_settings = '') {
-    $this->webroot = $app_root;
+  /**
+   * DrupalSettingsHandler constructor.
+   *
+   * @param string $path_to_webroot
+   *   The path to Drupal's webroot directory.
+   * @param string $database_key
+   *   Which key in the $databases array do you want to pull config from?
+   * @param string $path_to_settings
+   *   Absolute path to the settings file.
+   */
+  function __construct(string $path_to_webroot, string $database_key = 'default', string $path_to_settings = '') {
+    $this->webroot = $path_to_webroot;
     $this->databaseKey = $database_key;
     $this->pathToSettings = $path_to_settings;
   }
@@ -26,27 +36,12 @@ class DrupalSettingsHandler {
 
       // TODO Sniff for 9.
     }
-    if (!is_readable($this->pathToSettings)) {
-      throw new \RuntimeException(sprintf('%s settings file is not readable.', $this->pathToSettings));
-    }
-
-    if (!defined('DRUPAL_ROOT')) {
-      define('DRUPAL_ROOT', $this->webroot);
-    }
-    if (!defined('CONFIG_SYNC_DIRECTORY')) {
-      define('CONFIG_SYNC_DIRECTORY', 'sync');
-    }
-    require $this->pathToSettings;
-
-    global $db_url;
-    if (!empty($db_url)) {
-      return self::VERSION_6;
-    }
-
-    global $database;
-    if (!empty($database)) {
+    elseif (file_exists($this->webroot . '/scripts/drupal.sh')) {
       return self::VERSION_7;
     }
+
+    // TODO Handle Drupal 6?
+
   }
 
   public function getDatabaseConfig(int $drupal_version) {
@@ -61,6 +56,9 @@ class DrupalSettingsHandler {
    * @todo These have not been finished.
    */
   private function handleDrupal6() {
+
+    // TODO I think the drupal 7 code should work, just need a testcase, however this is so old may never get the chance. May 8, 2021 at 8:01:36 PM PDT, aklump.
+
     if (!is_readable($this->pathToSettings)) {
       throw new \RuntimeException(sprintf('%s settings file is not readable.', $this->pathToSettings));
     }
@@ -91,25 +89,12 @@ class DrupalSettingsHandler {
    * @todo These have not been finished.
    */
   private function handleDrupal7() {
-    if (!is_readable($this->pathToSettings)) {
-      throw new \RuntimeException(sprintf('%s settings file is not readable.', $this->pathToSettings));
-    }
-    if (!defined('DRUPAL_ROOT')) {
-      define('DRUPAL_ROOT', $this->webroot);
-    }
-    if (!defined('CONFIG_SYNC_DIRECTORY')) {
-      define('CONFIG_SYNC_DIRECTORY', 'sync');
-    }
-    require $this->pathToSettings;
-
+    define('DRUPAL_ROOT', $this->webroot);
+    require_once DRUPAL_ROOT . '/includes/bootstrap.inc';
+    drupal_bootstrap(DRUPAL_BOOTSTRAP_CONFIGURATION);
     global $databases;
 
-    return $databases[$this->databaseKey]['default'] + array_fill_keys(array(
-        'database',
-        'username',
-        'password',
-        'port',
-      ), NULL);
+    return $databases[$this->databaseKey]['default'] ?? [];
   }
 
   private function handleDrupal8() {
@@ -125,4 +110,5 @@ class DrupalSettingsHandler {
 
     return $databases[$this->databaseKey];
   }
+
 }